Move data field edition to mixin + fixes
