+++ src/devtools/components/StateInspector.vue
@@ -78,8 +78,7 @@ export default {
       })
     },
     dataTypeTooltip () {
-      const ctrlKey = navigator.platform === 'MacIntel' ? '&#8984;' : 'Ctrl'
-      return `<span class="keyboard">${ctrlKey}</span> + <i class="material-icons">mouse</i>: Collapse All<br><span class="keyboard">Shift</span> + <i class="material-icons">mouse</i>: Expand All`
+      return `<span class="keyboard">${this.$keys.ctrl}</span> + <i class="material-icons">mouse</i>: Collapse All<br><span class="keyboard">Shift</span> + <i class="material-icons">mouse</i>: Expand All`
     }
   },
   methods: {
